+++ local/metadata/session_form.php
@@ -1,0 +1,26 @@
+<?php
+require_once '../../config.php';
+require_once $CFG->dirroot.'/lib/formslib.php';
+require_once $CFG->dirroot.'/lib/datalib.php';
+
+class session_form extends moodleform {
+	function definition() {
+		global $CFG, $DB, $USER; //Declare our globals for use
+		$mform = $this->_form; //Tell this object to initialize with the properties of the Moodle form.
+
+		// Form elements
+		//$mform->addElement('text', 'email', get_string('email'));
+	}
+	
+	//If you need to validate your form information, you can override  the parent's validation method and write your own.	
+	function validation($data, $files) {
+		$errors = parent::validation($data, $files);
+		global $DB, $CFG, $USER; //Declare them if you need them
+
+		//if ($data['data_name'] Some condition here)  {
+		//	$errors['element_to_display_error'] = get_string('error', 'local_demo_plug-in');
+		//}
+	}
+}
+
+?>
